requirements.txt
Pillow>=10.0.0
PyYAML>=6.0
matplotlib>=3.7.0
